backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.35.2
torch>=2.2.0
numpy==1.26.2
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 